Revert "[Data] Update Export API metadata and refresh the dataset/operator state when there is a change (#54623)"
This reverts commit 7cb74e845b0a8edebed8d7e5d0c441474eab1a6e.
