+++ python/ray/data/_internal/metadata_exporter.py
@@ -13,7 +13,6 @@ from ray._private.event.export_event_logger import (
     check_export_api_enabled,
     get_export_event_logger,
 )
-from ray.data._internal.execution.dataset_state import DatasetState
 from ray.data.context import DataContext
 
 if TYPE_CHECKING:
@@ -61,17 +60,11 @@ class Operator:
         sub_stages: List of sub-stages contained within this operator.
         args: User-specified arguments associated with the operator, which may
             include configuration settings, options, or other relevant data for the operator.
-        execution_start_time: The timestamp when the operator execution begins.
-        execution_end_time: The timestamp when the operator execution ends.
-        state: The state of the operator.
     """
 
     name: str
     id: str
     uuid: str
-    execution_start_time: Optional[float]
-    execution_end_time: Optional[float]
-    state: str
     input_dependencies: List[str] = field(default_factory=list)
     sub_stages: List[SubStage] = field(default_factory=list)
     args: Dict[str, Any] = field(default_factory=dict)
@@ -115,9 +108,6 @@ class Topology:
                     op_to_id[dep] for dep in op.input_dependencies if dep in op_to_id
                 ],
                 args=sanitize_for_struct(op._get_logical_args()),
-                execution_start_time=None,
-                execution_end_time=None,
-                state=DatasetState.PENDING.name,
             )
 
             # Add sub-stages if they exist
@@ -141,11 +131,8 @@ class DatasetMetadata:
         job_id: The ID of the job running this dataset.
         topology: The structure of the dataset's operator DAG.
         dataset_id: The unique ID of the dataset.
-        start_time: The timestamp when the dataset is registered.
+        start_time: The timestamp when the dataset execution started.
         data_context: The DataContext attached to the dataset.
-        execution_start_time: The timestamp when the dataset execution starts.
-        execution_end_time: The timestamp when the dataset execution ends.
-        state: The state of the dataset.
     """
 
     job_id: str
@@ -153,9 +140,6 @@ class DatasetMetadata:
     dataset_id: str
     start_time: float
     data_context: DataContext
-    execution_start_time: Optional[float]
-    execution_end_time: Optional[float]
-    state: str
 
 
 def _add_ellipsis(s, truncate_length):
@@ -218,9 +202,6 @@ def dataset_metadata_to_proto(dataset_metadata: DatasetMetadata) -> Any:
             id=op.id,
             uuid=op.uuid,
             args=args,
-            execution_start_time=op.execution_start_time,
-            execution_end_time=op.execution_end_time,
-            state=ProtoOperator.OperatorState.Value(op.state),
         )
 
         # Add input dependencies
@@ -246,9 +227,6 @@ def dataset_metadata_to_proto(dataset_metadata: DatasetMetadata) -> Any:
         job_id=dataset_metadata.job_id,
         start_time=dataset_metadata.start_time,
         data_context=data_context,
-        execution_start_time=dataset_metadata.execution_start_time,
-        execution_end_time=dataset_metadata.execution_end_time,
-        state=ProtoDatasetMetadata.DatasetState.Value(dataset_metadata.state),
     )
     proto_dataset_metadata.topology.CopyFrom(proto_topology)
